Complete task 2 of Hello World
Completed task 2 of the project.
Corrections were also made to the "1-comiler" file. Previously, the ${CFILE}.o was adding the whole name of the C file to the output file, therefore prefixing the ".o" extension. This was changed to ${CFILE%.c}.o which subsequently removes the ".c" part of the file name before being concatenated to the ".o" extension to form the output file's name.