Make corrections to files failed by checker
All the files in this commit had a similar issue with the checker.
Prevoiusly, all of them had a for loop where the initial declaration of
a variable was done inside the loop; something that is not allowed in
older versions of C.
This was changed to a declaration of the said variable outside the for
loop.
There was no need of adding the varaible to the documentation of the
function.