Correct errors in tasks 0 and 3 of C Malloc
In task 3, there was an error caused by a betty styling issue. This was
caused by having spaces preceeding tab characters; something that was
prevalent from line 24 upto 38.
In task 0, excessive bytes were allocated for the operation. This was
caused by using "sizeof(int)" instead of "sizeof(char)", since we were
dealing with an array of characters.