Update files in task 5, 6, 9 and 10
Previously, the functions in these source files had the declaration of a
variable 'temp' that was being assigned the value of '*head'.
This is poor practice and could cause a segmentation fault in the case
where '*head' has a NULL reference. This was changed by moving this
assignment operation after checking whether or not '*head' has a NULL
reference.