Complete task 1 of File I/O
Corrections were also made to 0-read_textfile.c in an attempt to correct
failed check. Previously, the file descriptor was not closed in the
function. This was effectively corrected. Also, the size of the buffer
was also changed to a dynamically allocated memory block compared to the
static size of 1024 bytes, previously.